+++ src/Ant Simulator 2/obj.cpp
@@ -26,9 +26,9 @@ void obj::setRotation(float _a) {
 
 void obj::setRotation(float _dx, float _dy) {
 	if (_dx < 0)
-		this->setRotation((atan(_dy / _dx) * 180.f / M_PI - 90.f));
+		this->setRotation((atan(_dy / _dx) * 180.f / (float)M_PI - 90.f));
 	else
-		this->setRotation(atan(_dy / _dx) * 180.f / M_PI + 90.f);
+		this->setRotation(atan(_dy / _dx) * 180.f / (float)M_PI + 90.f);
 }
 
 float obj::getX() {
